backend/requirements.txt
python-dotenv
requests
flask
pyyaml